apps/s3-to-influxdb/requirements.txt
boto3~=1.24.75
jmespath~=1.0.1
influxdb-client~=1.33.0
PyYAML~=6.0